+++ okapi-operation/tests/axum_integration.rs
@@ -1,5 +1,49 @@
-#[cfg(feature = "axum-integration")]
-mod tests {
+#[cfg(feature = "axum")]
+#[allow(deprecated)]
+mod openapi {
+    use axum::Json;
+    use okapi::{openapi3::RefOr, schemars::gen::SchemaGenerator};
+    use okapi_operation::{
+        axum_integration::{get, Router},
+        oh, openapi, Components, ToMediaTypes, ToResponses,
+    };
+
+    #[test]
+    fn json_body_detection() {
+        #[openapi]
+        async fn handle(_arg: Json<String>) {}
+
+        let schema = Router::<()>::new()
+            .route("/", get(oh!(handle)))
+            .generate_openapi_builder()
+            .build()
+            .expect("Schema generation shoildn't fail");
+
+        let operation = schema.paths["/"]
+            .clone()
+            .get
+            .expect("GET / should be present")
+            .request_body
+            .expect("GET / request body should be present");
+        let RefOr::Object(request_body) = operation else {
+            panic!("GET / request body should be RefOr::Object");
+        };
+
+        let body_schema = request_body.content["application/json"]
+            .clone()
+            .schema
+            .expect("GET / request body schema should be present");
+
+        let mut gen = SchemaGenerator::default();
+        let expected_schema = gen.subschema_for::<String>().into_object();
+
+        assert_eq!(body_schema, expected_schema);
+    }
+}
+
+#[cfg(feature = "axum")]
+#[allow(deprecated)]
+mod openapi_handler {
     use axum::body::Body;
     use http::Request;
     use okapi_operation::{